Fixed warning in a DBG on Free-BSD

+++ modules_s/pike/pike_funcs.c
@@ -65,7 +65,7 @@ int pike_check_req(struct sip_msg *msg, char *foo, char *bar)
 	node = add_node( tree, msg->src_ip.u.addr,msg->src_ip.len,&father,&flag);
 
 	DBG("DEBUG:pike_check_req: src IP [%.*s]; hit node = [%d][%d] flags=%d\n",
-		msg->src_ip.len,msg->src_ip.u.addr,
+		(int)msg->src_ip.len,msg->src_ip.u.addr,
 		node->hits,node->leaf_hits,flag);
 
 	/* do all the job with the timer */
